requirements.txt
Flask==2.2.5
bidict==0.22.1
click==8.1.7
colorama==0.4.6
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
h11==0.14.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==2.1.0
pip==20.1.1
python-engineio==4.9.1
python-socketio==5.11.0
setuptools==47.1.0
simple-websocket==1.0.0
typing-extensions==4.7.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
